+++ src/views/TaskDetailView.ts
@@ -404,9 +404,7 @@ export class TaskDetailView extends ItemView {
         onChange: () => this.scheduleSave(),
         onPivot: (value) => void openIndicatorSearch(this.plugin, value),
         reputationKeys: {
-          virustotal: this.plugin.settings.virusTotalApiKey,
-          abuseipdb: this.plugin.settings.abuseIpdbApiKey,
-          abusech: this.plugin.settings.abuseChApiKey
+          ...this.plugin.reputationKeys()
         },
         findSightings: (value) => iocSightings(value, project.tasks, task.id)
       })